chore: keep it DRY
