+++ posthog/admin/admins/insight_admin.py
@@ -3,6 +3,7 @@ from django.contrib import admin
 from django.urls import reverse
 from django.utils.html import format_html
 
+from posthog.admin.filters import DeletedFilter
 from posthog.models import Insight
 
 
@@ -51,36 +52,6 @@ class InsightAdminForm(forms.ModelForm):
         return instance
 
 
-class DeletedInsightFilter(admin.SimpleListFilter):
-    title = "deleted"
-    parameter_name = "deleted"
-
-    def lookups(self, request, model_admin):
-        return (
-            ("no", "No"),
-            ("yes", "Yes"),
-            ("all", "All"),
-        )
-
-    def choices(self, changelist):
-        # Override default "All" choice so "No" is the default selected option
-        value = self.value() or "no"
-        for lookup, title in self.lookup_choices:
-            yield {
-                "selected": value == lookup,
-                "query_string": changelist.get_query_string({self.parameter_name: lookup}),
-                "display": title,
-            }
-
-    def queryset(self, request, queryset):
-        value = self.value() or "no"
-        if value == "no":
-            return queryset.filter(deleted=False)
-        if value == "yes":
-            return queryset.filter(deleted=True)
-        return queryset
-
-
 class InsightAdmin(admin.ModelAdmin):
     form = InsightAdminForm
     exclude = ("layouts",)
@@ -97,7 +68,7 @@ class InsightAdmin(admin.ModelAdmin):
         "deleted",
     )
     list_display_links = ("id", "short_id", "effective_name")
-    list_filter = (DeletedInsightFilter,)
+    list_filter = (DeletedFilter,)
     list_select_related = ("team", "team__organization")
     search_fields = ("id", "name", "short_id", "team__name", "team__organization__name")
     readonly_fields = (
